pyproject.toml
[tool.poetry]
name = "autonoumous car"
version = "0.1.0"
description = ""
authors = ["amyhsu <amy991225@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
opencv-python = "^4.10.0.84"
pyserial = "^3.5"
python-dotenv = "^1.0.1"
keyboard = "^0.13.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
